refactor(faro): isolate local whoami database probes
The remote client imported the root Duty package only to probe an optional local database and inspect access tokens. That linked database bootstrap machinery into every Faro build.

Expose the local checks through an injected interface and register the Duty-backed implementation only in the full Mission Control binary.
